combine ar and dmsg monitor as network-monitor
