+++ cmd/network-monitor/network-monitor.go
@@ -11,7 +11,7 @@ import (
 func init() {
 	var helpflag bool
 	commands.RootCmd.SetUsageTemplate(help)
-	commands.RootCmd.PersistentFlags().BoolVarP(&helpflag, "help", "h", false, "help for dmsgpty-cli")
+	commands.RootCmd.PersistentFlags().BoolVarP(&helpflag, "help", "h", false, "help for dmsg-monitor")
 	commands.RootCmd.SetHelpCommand(&cobra.Command{Hidden: true})
 	commands.RootCmd.PersistentFlags().MarkHidden("help") //nolint
 }
