--- pkg/network-monitor/store/memory_store_test.go
@@ -1,13 +1,0 @@
-//go:build !no_ci
-// +build !no_ci
-
-package store
-
-import (
-	"testing"
-)
-
-func TestMemory(t *testing.T) {
-	s := newMemoryStore()
-	testNetwork(t, s)
-}
